requirements.txt
# PyTorch with CUDA support (fallback to CPU if CUDA unavailable)
torch>=2.0.0
# For explicit CUDA 11.8 support, use:
# torch>=2.0.0+cu118 -f https://download.pytorch.org/whl/cu118/torch_stable.html
# For explicit CUDA 12.1 support, use:
# torch>=2.0.0+cu121 -f https://download.pytorch.org/whl/cu121/torch_stable.html
scikit-learn>=1.3.0
numpy>=1.24.0
pandas>=2.0.0
huggingface-hub>=0.20.0
datasets>=2.16.0
python-dotenv>=1.0.0
tqdm>=4.65.0
matplotlib>=3.7.0
seaborn>=0.12.0
PyGithub>=2.1.0
pytest>=7.4.0
pytest-cov>=4.1.0